refactor: pull image search logic into custom hook
